add new func

+++ ft_isascii.c
@@ -6,13 +6,13 @@
 /*   By: yoelhaim <yoelhaim@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 11:38:43 by yoelhaim          #+#    #+#             */
-/*   Updated: 2021/11/05 16:44:33 by yoelhaim         ###   ########.fr       */
+/*   Updated: 2021/11/06 11:22:48 by yoelhaim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_isascii(int c)
+int	ft_isascii(int c)
 {
 	if (c < 0 || c > 127)
 		return (0);
